api-flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
